pyproject.toml
[project]
name = "hierarchical-rl-agent-for-efficient-os-control"
version = "0.1.0"
description = "Hierarchical Reinforcement Learning Agent for Efficient OS Control - FastAPI Microservice"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Deep Learning - More Powerful Models
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.35.0",
    "timm>=0.9.12",
    "openai>=1.3.0",
    "clip @ git+https://github.com/openai/CLIP.git",
    
    # Reinforcement Learning
    "gymnasium>=0.29.1",
    "stable-baselines3>=2.2.1",
    "tensorboard>=2.15.0",
    
    # OSWorld Integration
    "osworld @ git+https://github.com/xlang-ai/OSWorld.git",
    "docker>=6.1.0",
    "requests>=2.31.0",
    
    # Computer Control
    "pyautogui>=0.9.54",
    "mss>=9.0.1",
    "pillow>=10.1.0",
    "opencv-python>=4.8.1",
    "pynput>=1.7.6",
    
    # GUI Visualization for Training
    "pygame>=2.5.0",
    "matplotlib>=3.8.0",
    
    # Utilities
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
